+++ Packages/com.yontalane.dialog/Runtime/DialogProcessor.cs
@@ -837,7 +837,7 @@ namespace Yontalane.Dialog
 
                 // Initialize the query UI with the description and responses, and set the OnQueryResponse callback
                 string description = !string.IsNullOrEmpty(queryData.description) ? queryData.description : string.Empty;
-                Query.QueryProcessor.InitiateWithDescription(ReplaceInlineText(queryData.text), ReplaceInlineText(description), responses, OnQueryResponse);
+                Query.QueryProcessor.InitiateWithDescription(ReplaceInlineText(queryData.text), ReplaceInlineText(description), responses, queryData.initialSelection, OnQueryResponse);
                 return true;
             }